added new QA Lab menu items as per tenant SG started dynamic class loading for Navigation methods
