+++ RKCIUIAutomation/Page/Navigation/Navigation.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RKCIUIAutomation.Page.Navigation
+{
+    public abstract class Navigation : PageBase, INavigation
+    {
+    }
+}
