+++ RKCIUIAutomation/Page/Navigation/Navigation_Impl.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RKCIUIAutomation.Page.Navigation
+{
+    public class Navigation_Impl : Navigation, INavigation
+    {
+    }
+}
