+++ RKCIUIAutomation/Page/PageObjects/RMCenter/Search/Search.cs
@@ -1,23 +1,60 @@
 ﻿using OpenQA.Selenium;
+using RKCIUIAutomation.Config;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using RKCIUIAutomation.Config;
-using static RKCIUIAutomation.Page.Action;
-using static RKCIUIAutomation.Page.PageHelper;
 
 namespace RKCIUIAutomation.Page.PageObjects.RMCenter.Search
 {
-    public class Search : Search_Impl, ISearch
+    public abstract class Search : PageBase, ISearch
     {
-        public Search() { }
+        private static ProjectName tenant;
 
-        public Search(IWebDriver driver) => Driver = driver;
-        public Search GetCommonInstance(IWebDriver driver) => new Search(driver);
+        //public Search() { }
+        //public Search(IWebDriver driver) => Driver = driver;
+        public static T SetClass<T>() => (T)SetPageClassBasedOnTenant();
 
-        private enum SearchCriteria
+        public abstract By GetPageTitleByLocator();
+        public abstract void EnterDate_From(string fromDate);
+        public abstract void EnterDate_To(string toDate);
+        public abstract void EnterText_Attention(string text);
+        public abstract void EnterText_From(string text);
+        public abstract void EnterText_MSLNumber(string text);
+        public abstract void EnterText_Number(string text);
+        public abstract void EnterText_OriginatorDocumentRef(string text);
+        public abstract void EnterText_Title(string text);
+        public abstract void EnterText_TransmittalNumber(string text);
+        public abstract void PopulateAllSearchCriteriaFields();
+        public abstract void SelectDDL_Category<T>(T itemIndexOrName);
+        public abstract void SelectDDL_DocumentType<T>(T itemIndexOrName);
+        public abstract void SelectDDL_SegmentArea<T>(T itemIndexOrName);
+        public abstract void SelectDDL_Status<T>(T itemIndexOrName);
+
+        private static object SetPageClassBasedOnTenant()
+        {
+            tenant = projectName;
+            
+            var instance = new Search_Impl();
+
+            if (tenant == ProjectName.GLX)
+            {
+                Console.WriteLine("###### using Search_GLX instance");
+                instance = new Search_GLX();
+            }
+            else if (tenant == ProjectName.I15Tech)
+            {
+                Console.WriteLine("###### using Search_I15Tech instance");
+                instance = new Search_I15Tech();
+            }
+            else
+                Console.WriteLine("###### using Search_Impl (Common) instance");
+
+            return instance;
+        }
+        
+        internal enum SearchCriteria
         {
             [StringValue("SelectedType")] DocumentType,
             [StringValue("SelectedStatus")] Status,
@@ -33,51 +70,5 @@ namespace RKCIUIAutomation.Page.PageObjects.RMCenter.Search
             [StringValue("TransmittalFromDate")] Transmittal_FromDate, //+GLX
             [StringValue("TransmittalToDate")] Transmittal_ToDate //+GLX
         }
-
-        public static By Txt_PageTitle { get; } = By.XPath("//h3");
-
-        //Page workflow common to all tenants
-        public override void SelectDDL_DocumentType<T>(T itemIndexOrName) => ExpandAndSelectFromDDList(SearchCriteria.DocumentType, itemIndexOrName);
-        public override void SelectDDL_Status<T>(T itemIndexOrName) => ExpandAndSelectFromDDList(SearchCriteria.Status, itemIndexOrName);
-        public override void EnterText_Title(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.Title), text);
-        public override void EnterText_TransmittalNumber(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.TransmittalNumber), text);
-        public override void EnterText_From(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.From), text);
-        public override void EnterText_Attention(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.Attention), text);
-
-        //Not used in tenant(s): GLX
-        public override void EnterText_Number(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.Number), text);
-        //Not used in tenant(s): GLX, I15Tech
-        public override void EnterText_MSLNumber(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.MSLNumber), text);
-
-
-        //Used only in GLX
-        public override void SelectDDL_Category<T>(T itemIndexOrName) => ExpandAndSelectFromDDList(SearchCriteria.Category, itemIndexOrName);
-        public override void SelectDDL_SegmentArea<T>(T itemIndexOrName) => ExpandAndSelectFromDDList(SearchCriteria.SegmentArea, itemIndexOrName);
-        public override void EnterText_OriginatorDocumentRef(string text) => EnterText(GetTextInputFieldByLocator(SearchCriteria.OriginatorDocumentRef), text);
-
-        /// <summary>
-        /// Date value in string format (i.e. MM/DD/YYYY)
-        /// </summary>
-        /// <param name="date"></param>
-        public override void EnterDate_From(string fromDate) => EnterText(GetTextInputFieldByLocator(SearchCriteria.Transmittal_FromDate), fromDate);
-
-        /// <summary>
-        /// Date value in string format (i.e. MM/DD/YYYY)
-        /// </summary>
-        /// <param name="toDate"></param>
-        public override void EnterDate_To(string toDate) => EnterText(GetTextInputFieldByLocator(SearchCriteria.Transmittal_ToDate), toDate);
-
-        public override void PopulateAllSearchCriteriaFields()
-        {
-            SelectDDL_DocumentType(1);
-            EnterText_Number("Common Test Number");
-            EnterText_TransmittalNumber("Common Test Transmittal Number");
-            EnterText_Title("Common Test Title");
-            EnterText_From("From Common Test");
-            EnterText_MSLNumber("Common Test MSL Number");
-            EnterText_Attention("Attention Common Test");
-        }
-
-        
     }
 }
