+++ RKCIUIAutomation/RKCIUIAutomation.csproj
@@ -229,6 +229,9 @@
     <Compile Include="Config\ProjectProperties.cs" />
     <Compile Include="Config\Properties.cs" />
     <Compile Include="Page\Action.cs" />
+    <Compile Include="Page\Navigation\INavigation.cs" />
+    <Compile Include="Page\Navigation\Navigation.cs" />
+    <Compile Include="Page\Navigation\Navigation_Impl.cs" />
     <Compile Include="Page\Navigation\NavMenu.cs" />
     <Compile Include="Page\PageObjects\IPageBase.cs" />
     <Compile Include="Page\PageObjects\Project\Administration\AdminTools\LockedRecords.cs" />
@@ -298,16 +301,14 @@
     <Compile Include="Page\PageObjects\RFI\List.cs" />
     <Compile Include="Page\PageObjects\RMCenter\DesignDocument.cs" />
     <Compile Include="Page\PageObjects\RMCenter\DOTProjectCorrespondenceLog.cs" />
-    <Compile Include="Page\PageObjects\RMCenter\Search\Search_Impl.cs" />
+    <Compile Include="Page\PageObjects\RMCenter\Search\Search.cs" />
     <Compile Include="Page\PageObjects\RMCenter\Search\ISearch.cs" />
     <Compile Include="Page\PageObjects\RMCenter\ProjectCorrespondenceLog.cs" />
     <Compile Include="Page\PageObjects\RMCenter\ReviewReviseSubmittal.cs" />
     <Compile Include="Page\PageObjects\RMCenter\RFCManagement.cs" />
     <Compile Include="Page\PageObjects\RMCenter\RMCenter.cs" />
     <Compile Include="Page\PageObjects\RMCenter\ReviewReviseSubmittals.cs" />
-    <Compile Include="Page\PageObjects\RMCenter\Search\Search.cs" />
-    <Compile Include="Page\PageObjects\RMCenter\Search\Search_GLX.cs" />
-    <Compile Include="Page\PageObjects\RMCenter\Search\Search_I15Tech.cs" />
+    <Compile Include="Page\PageObjects\RMCenter\Search\Search_Impl.cs" />
     <Compile Include="Page\PageObjects\RMCenter\SubmittalDetails.cs" />
     <Compile Include="Page\PageObjects\RMCenter\UploadDEVSubmittal.cs" />
     <Compile Include="Page\PageObjects\RMCenter\UploadOwnerSubmittal.cs" />
